Fix image comparison

+++ overlays/media-juggler-png-optimizer/media-juggler-png-optimizer.nu
@@ -15,6 +15,19 @@ export def image_data_hash []: [
   $result.stdout | str trim
 }
 
+# Compare two images with odiff
+export def same_image [
+  image2: path
+]: path -> bool {
+  let image1 = $in
+  let result = do {^odiff --threshold 0.0 $image1 $image2} | complete
+  if $result.exit_code not-in [0, 21, 22] {
+    log info $"Error running '^odiff --threshold 0.0 \"($image1)\" \"($image2)\"'\nstderr: ($result.stderr)\nstdout: ($result.stdout)"
+    return null
+  }
+  $result.exit_code == 0
+}
+
 # This is a wrapper program meant to be called by pdfsizeopt for optimizing images, specifically PNGs.
 #
 # It wraps a couple of other utilities, namely oxipng, ect, and pngcrush, and allows forcing images to gray scale.
@@ -38,11 +51,15 @@ def main [
     }
   )
 
-  let original_image_checksum = (
-    if not ($force_gray_scale) {
-      $source | image_data_hash
-    }
-  )
+  let extension = $source | path parse | get extension
+  let original_source = mktemp --suffix ("." + $extension) --tmpdir-path .
+  cp $source $original_source
+
+  # let original_image_checksum = (
+  #   if not ($force_gray_scale) {
+  #     $source | image_data_hash
+  #   }
+  # )
 
   # Use pngcrush to force images to gray scale.
   let source = (
@@ -79,27 +96,23 @@ def main [
   } | complete
   if $result.exit_code != 0 {
     log error $"Error running '^oxipng --opt max --out \"($destination)\" --quiet --strip safe -- \"($source)\"'\nstderr: ($result.stderr)\nstdout: ($result.stdout)"
+    rm --force $original_source
     exit 1
   }
-  if not ($force_gray_scale) and ($destination | path exists) {
-    let new_image_checksum = $destination | image_data_hash
-    if ($original_image_checksum != $new_image_checksum) {
-      log warning "oxipng produced an image with different image data. Rerunning without modifying colortype."
-      # Changes to the colortype may cause it to render differently, so redo without the colortype change.
-      let result = do {
-        ^oxipng --nc --opt max --out $destination --quiet --strip safe -- $source
-      } | complete
-      if $result.exit_code != 0 {
-        log error $"Error running '^oxipng --nc --opt max --out \"($destination)\" --quiet --strip safe -- \"($source)\"'\nstderr: ($result.stderr)\nstdout: ($result.stdout)"
-        exit 1
-      }
-      if ($destination | path exists) {
-        let new_image_checksum = $destination | image_data_hash
-        if ($original_image_checksum != $new_image_checksum) {
-          log error "oxipng produced an image with different image data even without modifying colortype! Skipping oxipng optimizations."
-          cp $source $destination
-        }
-      }
+  if not ($force_gray_scale) and ($destination | path exists) and not ($original_source | same_image $destination) {
+    log warning "oxipng produced an image with different image data. Rerunning without modifying colortype."
+    # Changes to the colortype may cause it to render differently, so redo without the colortype change.
+    let result = do {
+      ^oxipng --nc --opt max --out $destination --quiet --strip safe -- $source
+    } | complete
+    if $result.exit_code != 0 {
+      log error $"Error running '^oxipng --nc --opt max --out \"($destination)\" --quiet --strip safe -- \"($source)\"'\nstderr: ($result.stderr)\nstdout: ($result.stdout)"
+      rm --force $original_source
+      exit 1
+    }
+    if ($destination | path exists) and not ($original_source | same_image $destination) {
+      log error "oxipng produced an image with different image data even without modifying colortype! Skipping oxipng optimizations."
+      cp $source $destination
     }
   }
 
@@ -117,29 +130,26 @@ def main [
   } | complete
   if $result.exit_code != 0 {
     log error $"Error running '^ect -9 -strip --mt-deflate \"($temp_destination)\"'\nstderr: ($result.stderr)\nstdout: ($result.stdout)"
-    rm --force $temp_destination
+    rm --force $original_source $temp_destination
     exit 1
   }
-  if not ($force_gray_scale) {
-    let new_image_checksum = $temp_destination | image_data_hash
-    if ($original_image_checksum != $new_image_checksum) {
-      log warning "ect produced an image with different image data. Rerunning with the --reuse flag."
-      # Changes to the colortype cause it to render differently, so redo without the colortype change.
+  if not ($force_gray_scale) and not ($original_source | same_image $destination) {
+    log warning "ect produced an image with different image data. Rerunning with the --reuse flag."
+    # Changes to the colortype cause it to render differently, so redo without the colortype change.
+    cp --force $destination $temp_destination
+    let result = do {
+      ^ect -9 -strip --mt-deflate --reuse $temp_destination
+    } | complete
+    if $result.exit_code != 0 {
+      log error $"Error running '^ect -9 -strip --mt-deflate --reuse \"($temp_destination)\"'\nstderr: ($result.stderr)\nstdout: ($result.stdout)"
+      rm --force $original_source $temp_destination
+      exit 1
+    }
+    if not ($original_source | same_image $destination) {
+      log error "ect produced an image with different image data even with the --reuse flag. Ignoring"
       cp --force $destination $temp_destination
-      let result = do {
-        ^ect -9 -strip --mt-deflate --reuse $temp_destination
-      } | complete
-      if $result.exit_code != 0 {
-        log error $"Error running '^ect -9 -strip --mt-deflate --reuse \"($temp_destination)\"'\nstderr: ($result.stderr)\nstdout: ($result.stdout)"
-        rm --force $temp_destination
-        exit 1
-      }
-      let new_image_checksum = $temp_destination | image_data_hash
-      if ($original_image_checksum != $new_image_checksum) {
-        log error "ect produced an image with different image data even with the --reuse flag. Ignoring"
-        cp --force $destination $temp_destination
-      }
     }
   }
+  rm --force $original_source
   mv --force $temp_destination $destination
 }
